Add files via upload
Time Series Forecasting - Commuter Forecast
Prediction methods: Simple Average, Moving Average and Seasonal Autoregressive Integrated Moving Average
This is a dataset of high-speed rail service commuters, including ID, Datetime and Count. This data set is a data from 2012-08-25 to 2014-09-25. The fields include ID, Datetime and Commuter Count. 
The goal of this project is to divide the data into a training set and a test set, and use the training set to predict the number of commuters in the future. The test set is the prediction data corresponding to the real data, and the three prediction methods are compared and analyzed.